requirements.txt
wheel
gunicorn
gevent #required by gunicorn
Django==3.2.5
#django-payments==0.14.0
django-paypal==1.1.1
django-environ==0.4.5
sendgrid==6.6.0
pyotp==2.6.0
SpeechRecognition==3.8.1
pycryptodome==3.10.1
numpy==1.21.2
pandas==1.1.5
openpyxl==3.0.9
InvoiceGenerator==1.1.0
pyshorteners==1.0.1
razorpay==1.2.0
django-cities==0.6
django-timezone-field==4.2.1
django-formtools==2.3
django-multi-form-view==2.0.1
